Refactor statistics and types
Make Type::CloneType abstract. Change unsigned to size_t in some
places in namespace model. Add keyword override for overrided methods.
Refactor ToKeyValueStringPairs to ToKeyValueMap.Change CsvStats::GetMin
signature, remove GetMinMax_. Rename CsvStats::STDAndCentralMoment_ to
CsvStats::CalculateCentralMoment. Refactor CsvStats::ShowSample and
it's description. Add GetValue to TypedColumnData. Add IsOrdered static
method to Type. Add CreateColumnData static methods to FDAlgorithm.

+++ src/model/column_layout_relation_data.h
@@ -19,7 +19,7 @@ public:
 
     using RelationData::AbstractRelationData;
 
-    [[nodiscard]] unsigned int GetNumRows() const final {
+    [[nodiscard]] size_t GetNumRows() const final {
         if (column_data_.empty()) {
             return 0;
         }
